+++ 13-binary_tree_nodes.c
@@ -14,7 +14,7 @@
  */
 size_t binary_tree_nodes(const binary_tree_t *tree)
 {
-	unsigned int count = 0;
+	size_t count = 0;
 
 	if (tree)
 	{
